Assign Log Changes

+++ Android/Maintenance/app/src/main/java/com/android/maintenance/configuration/ConfigConstant.java
@@ -4,8 +4,8 @@ package com.android.maintenance.configuration;
  * Created by anand on 07-Sep-16.
  */
 public interface ConfigConstant {
-    public static int company_id=1;
-    //public static String url="http://10.250.200.58/MaintenanceWebService/api/";
+    public static Long company_id=1L;
+   // public static String url="http://10.250.200.58/MaintenanceWebService/api/";
     //192.168.0107
     public static String url="http://10.0.2.2:80/MaintenanceWebService/api/";
     public static String adminRole = "admin";
